Fix warning about missing assertions when using assert_api_conform in Minitest
