+++ lib/openapi_first/test/methods.rb
@@ -1,20 +1,25 @@
 # frozen_string_literal: true
 
+require_relative 'minitest_helpers'
+require_relative 'plain_helpers'
+
 module OpenapiFirst
+  # Test integration
   module Test
+    def self.minitest?(base)
+      base.include?(::Minitest::Assertions)
+    rescue NameError
+      false
+    end
+
     # Methods to use in integration tests
     module Methods
-      def assert_api_conform(status: nil, api: :default)
-        api = OpenapiFirst::Test[api]
-        request = respond_to?(:last_request) ? last_request : @request
-        response = respond_to?(:last_response) ? last_response : @response
-        if status && status != response.status
-          raise OpenapiFirst::Error,
-                "Expected status #{status}, but got #{response.status} " \
-                "from #{request.request_method.upcase} #{request.path}."
+      def self.included(base)
+        if Test.minitest?(base)
+          base.include(OpenapiFirst::Test::MinitestHelpers)
+        else
+          base.include(OpenapiFirst::Test::PlainHelpers)
         end
-        api.validate_request(request, raise_error: true)
-        api.validate_response(request, response, raise_error: true)
       end
     end
   end
